+++ prisma-filter-common/src/index.ts
@@ -1,2 +1,3 @@
 export * from './filter.enum';
 export * from './filter.interface';
+export * from './filter.builder';